Move get-bounding-client-rect polyfill into polyfills
